requirements.txt
azure-identity>=1.15.0
azure-kusto-data>=4.3.1
mcp>=1.0.0
python-dotenv>=1.0.0
